test(pipnn): adapt and colocate core tests
Use current main error APIs and keep private plus graph/config composition tests beside their owning implementation modules.
