+++ diskann/src/graph/pipnn/mod.rs
@@ -358,13 +358,11 @@ where
     T: VectorRepr + Send + Sync + 'static,
 {
     if data.nrows() == 0 {
-        return Err(ANNError::log_dimension_mismatch_error(
-            "PiPNN requires at least one data point".into(),
-        ));
+        return Err(ANNError::message("PiPNN requires at least one data point"));
     }
     if data.ncols() == 0 {
-        return Err(ANNError::log_dimension_mismatch_error(
-            "PiPNN requires at least one data dimension".into(),
+        return Err(ANNError::message(
+            "PiPNN requires at least one data dimension",
         ));
     }
     if data.nrows() > u32::MAX as usize {
@@ -374,7 +372,7 @@ where
         )));
     }
     data.nrows().checked_mul(data.ncols()).ok_or_else(|| {
-        ANNError::log_dimension_mismatch_error(format!(
+        ANNError::message(format!(
             "PiPNN dataset shape {} x {} overflows usize",
             data.nrows(),
             data.ncols()
@@ -390,7 +388,7 @@ where
     // parallel pass, and the complete partition allocation drops on return.
     let candidates = tracing::info_span!("pipnn.leaf_build").in_scope(|| {
         leaf_build::build_leaf_candidates(data, leaves, context.config.k, metric)
-            .map_err(ANNError::opaque)
+            .map_err(ANNError::new)
     })?;
     // Finalization consumes candidate lists and reuses their allocations for the
     // resulting adjacency where possible.
@@ -412,8 +410,382 @@ fn effective_metric<T: 'static>(metric: Metric) -> Metric {
 
 #[track_caller]
 fn config_error(message: impl std::fmt::Display) -> ANNError {
-    ANNError::log_index_config_error("PiPNN".into(), message.to_string())
+    ANNError::message(format!("PiPNN configuration: {message}"))
 }
 
 #[cfg(test)]
-mod tests;
+mod tests {
+    use super::*;
+    use half::f16;
+
+    #[test]
+    fn integer_normalized_cosine_uses_unnormalized_cosine() {
+        for metric in [
+            Metric::L2,
+            Metric::Cosine,
+            Metric::CosineNormalized,
+            Metric::InnerProduct,
+        ] {
+            let expected = if metric == Metric::CosineNormalized {
+                Metric::Cosine
+            } else {
+                metric
+            };
+            assert_eq!(effective_metric::<u8>(metric), expected);
+            assert_eq!(effective_metric::<i8>(metric), expected);
+            assert_eq!(effective_metric::<f32>(metric), metric);
+            assert_eq!(effective_metric::<f16>(metric), metric);
+        }
+    }
+}
+#[cfg(test)]
+#[allow(
+    clippy::expect_used,
+    clippy::unwrap_used,
+    reason = "deterministic test fixture construction must abort on invalid setup"
+)]
+mod build_graph_tests {
+    use super::{PiPNNBuildContext, PiPNNConfig, build_graph};
+    use crate::graph::config::{self, MaxDegree};
+    use diskann_utils::views::MatrixView;
+    use diskann_vector::distance::Metric;
+    use half::f16;
+    use rand::{Rng, SeedableRng, rngs::StdRng};
+
+    fn pipnn_config() -> PiPNNConfig {
+        PiPNNConfig {
+            c_max: 4,
+            c_min: 1,
+            p_samp: 0.5,
+            fanout: vec![2],
+            k: 1,
+            replicas: 1,
+        }
+    }
+
+    fn graph_config(metric: Metric, degree: usize) -> crate::graph::Config {
+        config::Builder::new_with(degree, MaxDegree::same(), 8, metric.into(), |builder| {
+            builder.alpha(1.2);
+        })
+        .build()
+        .unwrap()
+    }
+
+    fn pool(threads: usize) -> rayon::ThreadPool {
+        rayon::ThreadPoolBuilder::new()
+            .num_threads(threads)
+            .build()
+            .unwrap()
+    }
+
+    fn rows(graph: Vec<crate::graph::AdjacencyList<u32>>) -> Vec<Vec<u32>> {
+        graph.into_iter().map(Vec::from).collect()
+    }
+
+    fn assert_graph_invariants(
+        graph: &[crate::graph::AdjacencyList<u32>],
+        points: usize,
+        degree: usize,
+    ) {
+        assert_eq!(graph.len(), points);
+        for (source, row) in graph.iter().enumerate() {
+            assert!(row.len() <= degree);
+            let mut sorted = row.to_vec();
+            sorted.sort_unstable();
+            sorted.dedup();
+            assert_eq!(sorted.len(), row.len());
+            assert!(
+                row.iter()
+                    .all(|&id| (id as usize) < points && id as usize != source)
+            );
+        }
+    }
+
+    #[test]
+    fn builds_a_single_leaf_graph_for_real_dataset_ids() {
+        let data = [0.0_f32, 1.0, 2.0, 3.0];
+        let data = MatrixView::try_from(&data[..], 4, 1).unwrap();
+        let graph = graph_config(Metric::L2, 2);
+        let pool = pool(2);
+        let context = PiPNNBuildContext::new(pipnn_config(), &graph, Metric::L2, &pool).unwrap();
+
+        let actual = build_graph(data, &context).unwrap();
+
+        assert_eq!(rows(actual), [vec![1], vec![0, 2], vec![1, 3], vec![2]]);
+
+        let graph = graph_config(Metric::L2, 1);
+        let context = PiPNNBuildContext::new(pipnn_config(), &graph, Metric::L2, &pool).unwrap();
+
+        let pruned = build_graph(data, &context).unwrap();
+
+        assert_graph_invariants(&pruned, 4, 1);
+        for (source, neighbors) in pruned.iter().enumerate() {
+            assert_eq!(source.abs_diff(neighbors[0] as usize), 1);
+        }
+    }
+
+    #[test]
+    fn prunes_complete_single_leaf_candidates_to_the_graph_degree() {
+        let data = [0.0_f32, 1.0, 2.0, 3.0, 4.0];
+        let data = MatrixView::try_from(&data[..], 5, 1).unwrap();
+        let graph = graph_config(Metric::L2, 1);
+        let pool = pool(2);
+        let config = PiPNNConfig {
+            c_max: 5,
+            c_min: 1,
+            p_samp: 0.5,
+            fanout: vec![2],
+            k: 4,
+            replicas: 1,
+        };
+        let context = PiPNNBuildContext::new(config, &graph, Metric::L2, &pool).unwrap();
+
+        let actual = build_graph(data, &context).unwrap();
+
+        assert_graph_invariants(&actual, 5, 1);
+        assert!(actual.iter().all(|row| row.len() == 1));
+    }
+
+    #[test]
+    fn rejects_empty_dataset_dimensions_at_the_public_boundary() {
+        let graph = graph_config(Metric::L2, 2);
+        let pool = pool(1);
+        let context = PiPNNBuildContext::new(pipnn_config(), &graph, Metric::L2, &pool).unwrap();
+
+        let no_rows = MatrixView::try_from(&[] as &[f32], 0, 4).unwrap();
+        let no_columns = MatrixView::try_from(&[] as &[f32], 4, 0).unwrap();
+
+        assert!(build_graph(no_rows, &context).is_err());
+        assert!(build_graph(no_columns, &context).is_err());
+    }
+
+    #[test]
+    fn supports_every_source_type_and_metric() {
+        fn build<T: crate::utils::VectorRepr + Send + Sync + 'static>(
+            values: &[T],
+            metric: Metric,
+        ) {
+            let data = MatrixView::try_from(values, 6, 2).unwrap();
+            let graph = graph_config(metric, 2);
+            let pool = pool(2);
+            let context = PiPNNBuildContext::new(pipnn_config(), &graph, metric, &pool).unwrap();
+            let actual = build_graph(data, &context).unwrap();
+            assert_graph_invariants(&actual, 6, 2);
+        }
+
+        let values = [
+            1.0_f32, 0.0, 0.0, 1.0, -1.0, 0.0, 0.0, -1.0, 0.5, 0.5, -0.5, -0.5,
+        ];
+        for metric in [
+            Metric::L2,
+            Metric::Cosine,
+            Metric::CosineNormalized,
+            Metric::InnerProduct,
+        ] {
+            build(&values, metric);
+        }
+        build(&values.map(f16::from_f32), Metric::L2);
+        build(&[1_u8, 0, 0, 1, 2, 0, 0, 2, 1, 1, 2, 2], Metric::L2);
+        build(&[1_i8, 0, 0, 1, -1, 0, 0, -1, 1, 1, -1, -1], Metric::L2);
+    }
+
+    #[test]
+    fn integer_normalized_cosine_matches_cosine() {
+        fn assert_match<T: crate::utils::VectorRepr + Send + Sync + 'static>(values: &[T]) {
+            let data = MatrixView::try_from(values, 8, 2).unwrap();
+            let pool = pool(2);
+            let build = |metric| {
+                let graph = graph_config(metric, 2);
+                let config = PiPNNConfig {
+                    c_max: 8,
+                    c_min: 1,
+                    p_samp: 0.5,
+                    fanout: vec![2],
+                    k: 1,
+                    replicas: 1,
+                };
+                let context = PiPNNBuildContext::new(config, &graph, metric, &pool).unwrap();
+                rows(build_graph(data, &context).unwrap())
+            };
+            assert_eq!(build(Metric::CosineNormalized), build(Metric::Cosine));
+        }
+
+        assert_match(&[1_u8, 0, 100, 1, 2, 0, 0, 1, 1, 1, 200, 2, 2, 1, 1, 2]);
+        assert_match(&[1_i8, 0, 100, 1, 2, 0, 0, 1, 1, 1, 120, 2, 2, 1, 1, 2]);
+    }
+
+    #[test]
+    fn is_deterministic_for_a_fixed_pool_size() {
+        let data: Vec<f32> = (0..96 * 4)
+            .map(|value| ((value * 17 + 3) % 101) as f32)
+            .collect();
+        let data = MatrixView::try_from(&data[..], 96, 4).unwrap();
+        let graph = graph_config(Metric::L2, 8);
+        let pool = pool(4);
+        let config = PiPNNConfig {
+            c_max: 16,
+            c_min: 4,
+            p_samp: 0.25,
+            fanout: vec![3, 2],
+            k: 3,
+            replicas: 2,
+        };
+        let context = PiPNNBuildContext::new(config, &graph, Metric::L2, &pool).unwrap();
+
+        let first = build_graph(data, &context).unwrap();
+        let second = build_graph(data, &context).unwrap();
+
+        assert_eq!(first, second);
+        assert_graph_invariants(&first, 96, 8);
+    }
+
+    #[test]
+    fn fixed_seed_randomized_sweeps_preserve_graph_invariants() {
+        let mut rng = StdRng::seed_from_u64(0x857a_d38b_44c2_0f11);
+        for case in 0..24 {
+            let points = rng.random_range(4..=32);
+            let dimensions = rng.random_range(1..=8);
+            let c_max = rng.random_range(4..=points.min(12));
+            let c_min = rng.random_range(1..=c_max);
+            let degree = rng.random_range(1..=points.min(8));
+            let values: Vec<f32> = (0..points * dimensions)
+                .map(|_| rng.random_range(-10.0..10.0))
+                .collect();
+            let data = MatrixView::try_from(&values[..], points, dimensions).unwrap();
+            let graph = graph_config(Metric::L2, degree);
+            let pool = pool(2);
+            let config = PiPNNConfig {
+                c_max,
+                c_min,
+                p_samp: 0.5,
+                fanout: vec![2],
+                k: rng.random_range(1..=3),
+                replicas: rng.random_range(1..=2),
+            };
+            let context = PiPNNBuildContext::new(config, &graph, Metric::L2, &pool).unwrap();
+
+            let actual = build_graph(data, &context)
+                .unwrap_or_else(|error| panic!("randomized case {case} failed: {error}"));
+            assert_graph_invariants(&actual, points, degree);
+        }
+    }
+}
+#[cfg(test)]
+#[allow(
+    clippy::expect_used,
+    clippy::unwrap_used,
+    reason = "deterministic test fixture construction must abort on invalid setup"
+)]
+mod config_tests {
+    use super::{PiPNNBuildContext, PiPNNConfig};
+    use crate::graph::config::{self, MaxDegree};
+    use diskann_vector::distance::Metric;
+
+    fn pipnn_config() -> PiPNNConfig {
+        PiPNNConfig {
+            c_max: 512,
+            c_min: 64,
+            p_samp: 0.01,
+            fanout: vec![10, 3],
+            k: 2,
+            replicas: 1,
+        }
+    }
+
+    fn graph_config(metric: Metric, alpha: f32) -> crate::graph::Config {
+        config::Builder::new_with(64, MaxDegree::same(), 72, metric.into(), |builder| {
+            builder.alpha(alpha);
+        })
+        .build()
+        .unwrap()
+    }
+
+    fn pool() -> rayon::ThreadPool {
+        rayon::ThreadPoolBuilder::new()
+            .num_threads(2)
+            .build()
+            .unwrap()
+    }
+
+    #[test]
+    fn rejects_each_invalid_algorithm_parameter() {
+        let graph = graph_config(Metric::L2, 1.2);
+        let pool = pool();
+        let mut cases = [
+            PiPNNConfig {
+                c_max: 0,
+                ..pipnn_config()
+            },
+            PiPNNConfig {
+                c_min: 0,
+                ..pipnn_config()
+            },
+            PiPNNConfig {
+                c_min: 513,
+                ..pipnn_config()
+            },
+            PiPNNConfig {
+                p_samp: 0.0,
+                ..pipnn_config()
+            },
+            PiPNNConfig {
+                p_samp: -0.01,
+                ..pipnn_config()
+            },
+            PiPNNConfig {
+                p_samp: 1.01,
+                ..pipnn_config()
+            },
+            PiPNNConfig {
+                p_samp: f64::NAN,
+                ..pipnn_config()
+            },
+            PiPNNConfig {
+                fanout: Vec::new(),
+                ..pipnn_config()
+            },
+            PiPNNConfig {
+                fanout: vec![1, 0],
+                ..pipnn_config()
+            },
+            PiPNNConfig {
+                fanout: vec![17],
+                ..pipnn_config()
+            },
+            PiPNNConfig {
+                k: 0,
+                ..pipnn_config()
+            },
+            PiPNNConfig {
+                replicas: 0,
+                ..pipnn_config()
+            },
+        ];
+
+        for config in &mut cases {
+            let error = PiPNNBuildContext::new(config.clone(), &graph, Metric::L2, &pool)
+                .expect_err("invalid PiPNN config must be rejected");
+            assert_eq!(error.kind(), diskann::ANNErrorKind::IndexConfigError);
+        }
+    }
+
+    #[test]
+    fn rejects_graph_policy_for_a_different_metric() {
+        let graph = graph_config(Metric::InnerProduct, 1.2);
+        let pool = pool();
+
+        let error = PiPNNBuildContext::new(pipnn_config(), &graph, Metric::L2, &pool).unwrap_err();
+
+        assert_eq!(error.kind(), diskann::ANNErrorKind::IndexConfigError);
+        assert!(error.to_string().contains("prune kind"));
+    }
+
+    #[test]
+    fn does_not_add_alpha_validation_beyond_graph_config() {
+        let pool = pool();
+        for alpha in [0.9, f32::NAN, f32::INFINITY] {
+            let graph = graph_config(Metric::L2, alpha);
+            PiPNNBuildContext::new(pipnn_config(), &graph, Metric::L2, &pool).unwrap();
+        }
+    }
+}
