+++ diskann/src/graph/pipnn/partitioning.rs
@@ -149,13 +149,13 @@ where
 {
     let points = data.nrows();
     if points == 0 {
-        return Err(ANNError::opaque(PartitionError::EmptyDataset));
+        return Err(ANNError::new(PartitionError::EmptyDataset));
     }
     if data.ncols() == 0 {
-        return Err(ANNError::opaque(PartitionError::EmptyDimensions));
+        return Err(ANNError::new(PartitionError::EmptyDimensions));
     }
     if points > u32::MAX as usize {
-        return Err(ANNError::opaque(PartitionError::TooManyPoints(points)));
+        return Err(ANNError::new(PartitionError::TooManyPoints(points)));
     }
 
     let mut leaves = Vec::new();
@@ -170,7 +170,7 @@ where
             partition_replica(data, &config, metric, &kernel, seed, &stripe_buffers)?;
         leaves
             .try_reserve(replica_leaves.len())
-            .map_err(ANNError::opaque)?;
+            .map_err(ANNError::new)?;
         leaves.append(&mut replica_leaves);
     }
     validate_leaves(&leaves, config.c_max)?;
@@ -191,14 +191,14 @@ where
     let initial_indices = point_ids(data.nrows())?;
     if data.nrows() <= config.c_max {
         let mut leaves = Vec::new();
-        leaves.try_reserve_exact(1).map_err(ANNError::opaque)?;
+        leaves.try_reserve_exact(1).map_err(ANNError::new)?;
         leaves.push(initial_indices);
         return Ok(leaves);
     }
 
     let mut leaves = Vec::new();
     let mut work = Vec::new();
-    work.try_reserve_exact(1).map_err(ANNError::opaque)?;
+    work.try_reserve_exact(1).map_err(ANNError::new)?;
     work.push(WorkItem {
         indices: initial_indices,
         level: 0,
@@ -213,7 +213,7 @@ where
         let mut results = Vec::new();
         results
             .try_reserve_exact(work.len())
-            .map_err(ANNError::opaque)?;
+            .map_err(ANNError::new)?;
         results.resize_with(work.len(), || None);
         // build_graph installs this complete private call tree into the
         // caller-owned pool; the indexed fill cannot escape that pool.
@@ -235,13 +235,11 @@ where
         let mut next_work = Vec::new();
         for result in results {
             let (mut pending, mut finished) =
-                result.ok_or_else(|| ANNError::opaque(PartitionError::MissingWorkerResult))??;
+                result.ok_or_else(|| ANNError::new(PartitionError::MissingWorkerResult))??;
             next_work
                 .try_reserve(pending.len())
-                .map_err(ANNError::opaque)?;
-            leaves
-                .try_reserve(finished.len())
-                .map_err(ANNError::opaque)?;
+                .map_err(ANNError::new)?;
+            leaves.try_reserve(finished.len()).map_err(ANNError::new)?;
             next_work.append(&mut pending);
             leaves.append(&mut finished);
         }
@@ -254,7 +252,7 @@ where
     let Some(largest) = work.iter().max_by_key(|item| item.indices.len()) else {
         return global_merge_small(leaves, config.c_min, config.c_max);
     };
-    Err(ANNError::opaque(PartitionError::IterationLimit {
+    Err(ANNError::new(PartitionError::IterationLimit {
         size: largest.indices.len(),
         level: largest.level,
         limit: MAX_PARTITION_ITERATIONS,
@@ -291,12 +289,10 @@ where
 
     let mut pending = Vec::new();
     let mut finished = Vec::new();
-    pending
-        .try_reserve(clusters.len())
-        .map_err(ANNError::opaque)?;
+    pending.try_reserve(clusters.len()).map_err(ANNError::new)?;
     finished
         .try_reserve(clusters.len())
-        .map_err(ANNError::opaque)?;
+        .map_err(ANNError::new)?;
     let child_seed = mix_seed(item.seed, points as u64);
     for cluster in clusters {
         if cluster.is_empty() {
@@ -319,7 +315,7 @@ fn sample_leaders(points: &[u32], sampling_fraction: f64, seed: u64) -> ANNResul
     let count = sample_num_leaders(points.len(), sampling_fraction);
     let mut rng = rand::rngs::StdRng::seed_from_u64(seed);
     let mut leaders = Vec::new();
-    leaders.try_reserve_exact(count).map_err(ANNError::opaque)?;
+    leaders.try_reserve_exact(count).map_err(ANNError::new)?;
     leaders.extend(points.choose_multiple(&mut rng, count).copied());
     Ok(leaders)
 }
@@ -478,7 +474,7 @@ where
         None,
         dots,
     )
-    .map_err(ANNError::opaque)?;
+    .map_err(ANNError::new)?;
 
     let point_scales = if metric == Metric::Cosine {
         grow_fallible(point_scale_buffer, point_count, 0.0)?;
@@ -504,14 +500,14 @@ where
         Metric::CosineNormalized | Metric::InnerProduct => PartitionScales::None,
     };
     let dots = MatrixView::try_from(&*dots, point_count, leader_count).map_err(|_| {
-        ANNError::opaque(PartitionError::InvalidBufferLength {
+        ANNError::new(PartitionError::InvalidBufferLength {
             buffer: "dot-product stripe",
             expected: dots_len,
             actual: dots.len(),
         })
     })?;
     let output = MutMatrixView::try_from(assignments, point_count, fanout).map_err(|error| {
-        ANNError::opaque(PartitionError::InvalidBufferLength {
+        ANNError::new(PartitionError::InvalidBufferLength {
             buffer: "partition assignments",
             expected: output_len,
             actual: error.into_inner().len(),
@@ -519,7 +515,7 @@ where
     })?;
     kernel
         .nearest_leaders(PartitionInput { dots, scales }, output)
-        .map_err(ANNError::opaque)
+        .map_err(ANNError::new)
 }
 
 fn gather_vectors<T>(data: MatrixView<'_, T>, indices: &[u32], output: &mut [f32]) -> ANNResult<()>
@@ -528,7 +524,7 @@ where
 {
     let expected = checked_area("gather output", indices.len(), data.ncols())?;
     if output.len() != expected {
-        return Err(ANNError::opaque(PartitionError::InvalidBufferLength {
+        return Err(ANNError::new(PartitionError::InvalidBufferLength {
             buffer: "gather output",
             expected,
             actual: output.len(),
@@ -560,9 +556,7 @@ fn scatter_assignments(
     let stripe_assignment_count = checked_area("scatter assignment stripe", stripe_points, fanout)?;
     let stripes = points.len().div_ceil(stripe_points);
     let mut partials = Vec::new();
-    partials
-        .try_reserve_exact(stripes)
-        .map_err(ANNError::opaque)?;
+    partials.try_reserve_exact(stripes).map_err(ANNError::new)?;
     partials.resize_with(stripes, || None);
     // See the pool invariant at the other partition terminal operations.
     #[allow(clippy::disallowed_methods)]
@@ -578,18 +572,16 @@ fn scatter_assignments(
         });
 
     let mut locals = Vec::new();
-    locals
-        .try_reserve_exact(stripes)
-        .map_err(ANNError::opaque)?;
+    locals.try_reserve_exact(stripes).map_err(ANNError::new)?;
     for result in partials {
-        locals.push(result.ok_or_else(|| ANNError::opaque(PartitionError::MissingWorkerResult))??);
+        locals.push(result.ok_or_else(|| ANNError::new(PartitionError::MissingWorkerResult))??);
     }
 
     let mut sizes = filled_vec(leaders, 0usize)?;
     for local in &locals {
         for (size, cluster) in sizes.iter_mut().zip(local) {
             *size = size.checked_add(cluster.len()).ok_or_else(|| {
-                ANNError::opaque(PartitionError::ShapeOverflow {
+                ANNError::new(PartitionError::ShapeOverflow {
                     buffer: "cluster size",
                     rows: *size,
                     cols: cluster.len(),
@@ -605,7 +597,7 @@ fn scatter_assignments(
         .enumerate()
         .map(|(leader, size)| {
             let mut cluster = Vec::new();
-            cluster.try_reserve_exact(size).map_err(ANNError::opaque)?;
+            cluster.try_reserve_exact(size).map_err(ANNError::new)?;
             for local in &locals {
                 cluster.extend_from_slice(&local[leader]);
             }
@@ -623,14 +615,14 @@ fn scatter_serial(
     let mut sizes = filled_vec(leaders, 0usize)?;
     for &leader in assignments {
         let Some(size) = sizes.get_mut(leader as usize) else {
-            return Err(ANNError::opaque(PartitionError::InvalidBufferLength {
+            return Err(ANNError::new(PartitionError::InvalidBufferLength {
                 buffer: "leader assignment",
                 expected: leaders,
                 actual: leader as usize + 1,
             }));
         };
         *size = size.checked_add(1).ok_or_else(|| {
-            ANNError::opaque(PartitionError::ShapeOverflow {
+            ANNError::new(PartitionError::ShapeOverflow {
                 buffer: "cluster size",
                 rows: *size,
                 cols: 1,
@@ -650,10 +642,10 @@ fn clusters_with_capacities(sizes: &[usize]) -> ANNResult<Vec<Vec<u32>>> {
     let mut clusters = Vec::new();
     clusters
         .try_reserve_exact(sizes.len())
-        .map_err(ANNError::opaque)?;
+        .map_err(ANNError::new)?;
     for &size in sizes {
         let mut cluster = Vec::new();
-        cluster.try_reserve_exact(size).map_err(ANNError::opaque)?;
+        cluster.try_reserve_exact(size).map_err(ANNError::new)?;
         clusters.push(cluster);
     }
     Ok(clusters)
@@ -666,10 +658,10 @@ fn global_merge_small(
 ) -> ANNResult<Vec<Vec<u32>>> {
     let mut merged = Vec::new();
     let mut small_leaves = Vec::new();
-    merged.try_reserve(leaves.len()).map_err(ANNError::opaque)?;
+    merged.try_reserve(leaves.len()).map_err(ANNError::new)?;
     small_leaves
         .try_reserve(leaves.len())
-        .map_err(ANNError::opaque)?;
+        .map_err(ANNError::new)?;
     for leaf in leaves {
         if leaf.len() >= c_min {
             merged.push(leaf);
@@ -682,11 +674,11 @@ fn global_merge_small(
     }
 
     let mut small = HashSet::new();
-    small.try_reserve(c_max).map_err(ANNError::opaque)?;
+    small.try_reserve(c_max).map_err(ANNError::new)?;
 
     for leaf in small_leaves {
         let combined = small.len().checked_add(leaf.len()).ok_or_else(|| {
-            ANNError::opaque(PartitionError::ShapeOverflow {
+            ANNError::new(PartitionError::ShapeOverflow {
                 buffer: "small-leaf merge",
                 rows: small.len(),
                 cols: leaf.len(),
@@ -695,7 +687,7 @@ fn global_merge_small(
         if combined > c_max {
             merged.push(drain_sorted(&mut small)?);
         }
-        small.try_reserve(leaf.len()).map_err(ANNError::opaque)?;
+        small.try_reserve(leaf.len()).map_err(ANNError::new)?;
         small.extend(leaf);
         if small.len() >= c_min {
             merged.push(drain_sorted(&mut small)?);
@@ -709,15 +701,14 @@ fn global_merge_small(
         {
             remainder.retain(|id| !last.contains(id));
             let combined = last.len().checked_add(remainder.len()).ok_or_else(|| {
-                ANNError::opaque(PartitionError::ShapeOverflow {
+                ANNError::new(PartitionError::ShapeOverflow {
                     buffer: "small-leaf tail merge",
                     rows: last.len(),
                     cols: remainder.len(),
                 })
             })?;
             if combined <= c_max {
-                last.try_reserve(remainder.len())
-                    .map_err(ANNError::opaque)?;
+                last.try_reserve(remainder.len()).map_err(ANNError::new)?;
                 last.append(&mut remainder);
                 last.sort_unstable();
             }
@@ -733,9 +724,7 @@ fn global_merge_small(
 
 fn drain_sorted(set: &mut HashSet<u32>) -> ANNResult<Vec<u32>> {
     let mut values = Vec::new();
-    values
-        .try_reserve_exact(set.len())
-        .map_err(ANNError::opaque)?;
+    values.try_reserve_exact(set.len()).map_err(ANNError::new)?;
     values.extend(set.drain());
     values.sort_unstable();
     Ok(values)
@@ -746,7 +735,7 @@ fn validate_leaves(leaves: &[Vec<u32>], c_max: usize) -> ANNResult<()> {
         .iter()
         .find(|leaf| leaf.is_empty() || leaf.len() > c_max)
     {
-        return Err(ANNError::opaque(PartitionError::InvalidLeaf {
+        return Err(ANNError::new(PartitionError::InvalidLeaf {
             size: leaf.len(),
             limit: c_max,
         }));
@@ -756,14 +745,14 @@ fn validate_leaves(leaves: &[Vec<u32>], c_max: usize) -> ANNResult<()> {
 
 fn point_ids(points: usize) -> ANNResult<Vec<u32>> {
     let mut ids = Vec::new();
-    ids.try_reserve_exact(points).map_err(ANNError::opaque)?;
+    ids.try_reserve_exact(points).map_err(ANNError::new)?;
     ids.extend(0..points as u32);
     Ok(ids)
 }
 
 fn filled_vec<T: Clone>(len: usize, value: T) -> ANNResult<Vec<T>> {
     let mut values = Vec::new();
-    values.try_reserve_exact(len).map_err(ANNError::opaque)?;
+    values.try_reserve_exact(len).map_err(ANNError::new)?;
     values.resize(len, value);
     Ok(values)
 }
@@ -779,14 +768,14 @@ fn grow_fallible<T: Clone>(values: &mut Vec<T>, len: usize, value: T) -> ANNResu
     }
     values
         .try_reserve(len - values.len())
-        .map_err(ANNError::opaque)?;
+        .map_err(ANNError::new)?;
     values.resize(len, value);
     Ok(())
 }
 
 fn checked_area(buffer: &'static str, rows: usize, cols: usize) -> ANNResult<usize> {
     rows.checked_mul(cols)
-        .ok_or_else(|| ANNError::opaque(PartitionError::ShapeOverflow { buffer, rows, cols }))
+        .ok_or_else(|| ANNError::new(PartitionError::ShapeOverflow { buffer, rows, cols }))
 }
 
 fn assignment_stripe_point_count(leader_count: usize) -> usize {
@@ -800,4 +789,423 @@ fn assignment_stripe_point_count(leader_count: usize) -> usize {
 }
 
 #[cfg(test)]
-mod tests;
+mod tests {
+    use diskann_utils::views::{Matrix, MatrixView};
+    use diskann_vector::{Half, distance::Metric};
+
+    use super::*;
+
+    fn config(c_min: usize, c_max: usize, fanout: Vec<usize>, replicas: usize) -> PiPNNConfig {
+        PiPNNConfig {
+            c_max,
+            c_min,
+            p_samp: 0.25,
+            fanout,
+            k: 1,
+            replicas,
+        }
+    }
+
+    fn clustered_data(points: usize, dimensions: usize) -> Matrix<f32> {
+        Matrix::new(
+            diskann_utils::views::Init({
+                let mut position = 0usize;
+                move || {
+                    let point = position / dimensions;
+                    let dimension = position % dimensions;
+                    position += 1;
+                    (point / 8) as f32 * 10.0 + dimension as f32 * 0.01 + point as f32 * 0.001
+                }
+            }),
+            points,
+            dimensions,
+        )
+    }
+
+    fn directional_data(points: usize, dimensions: usize) -> Matrix<f32> {
+        Matrix::new(
+            diskann_utils::views::Init({
+                let mut position = 0usize;
+                move || {
+                    let point = position / dimensions;
+                    let dimension = position % dimensions;
+                    position += 1;
+                    let angle = std::f32::consts::TAU * point as f32 / points as f32;
+                    match dimension {
+                        0 => angle.cos(),
+                        1 => angle.sin(),
+                        _ => 0.0,
+                    }
+                }
+            }),
+            points,
+            dimensions,
+        )
+    }
+
+    fn sorted_memberships(leaves: &[Vec<u32>]) -> Vec<Vec<u32>> {
+        let mut memberships: Vec<Vec<u32>> = leaves
+            .iter()
+            .map(|leaf| {
+                let mut ids = leaf.clone();
+                ids.sort_unstable();
+                ids
+            })
+            .collect();
+        memberships.sort();
+        memberships
+    }
+
+    fn assert_valid_partition(leaves: &[Vec<u32>], points: usize, c_max: usize, replicas: usize) {
+        assert!(
+            leaves
+                .iter()
+                .all(|leaf| !leaf.is_empty() && leaf.len() <= c_max)
+        );
+        let mut counts = vec![0usize; points];
+        for leaf in leaves {
+            let mut ids = leaf.clone();
+            ids.sort_unstable();
+            ids.dedup();
+            assert_eq!(ids.len(), leaf.len(), "duplicate ID inside a leaf");
+            for &id in leaf {
+                assert!((id as usize) < points);
+                counts[id as usize] += 1;
+            }
+        }
+        assert!(counts.iter().all(|&count| count >= replicas));
+    }
+
+    #[test]
+    fn returns_one_leaf_at_and_below_c_max() {
+        for points in [7, 8] {
+            let data = clustered_data(points, 3);
+            let leaves = partition(data.as_view(), config(2, 8, vec![2], 1), Metric::L2).unwrap();
+            assert_eq!(leaves, vec![(0..points as u32).collect::<Vec<_>>()]);
+        }
+    }
+
+    #[test]
+    fn partition_is_fixed_seed_deterministic_and_bounded() {
+        let data = clustered_data(96, 8);
+        let config = config(4, 16, vec![3, 2], 2);
+
+        let first = partition(data.as_view(), config.clone(), Metric::L2).unwrap();
+        let second = partition(data.as_view(), config, Metric::L2).unwrap();
+
+        assert_eq!(sorted_memberships(&first), sorted_memberships(&second));
+        assert_valid_partition(&first, 96, 16, 2);
+        assert!(first.iter().map(Vec::len).sum::<usize>() > 96 * 2);
+    }
+
+    #[test]
+    fn partition_remains_bounded_after_the_fanout_schedule_is_exhausted() {
+        let data = clustered_data(80, 4);
+        let leaves = partition(data.as_view(), config(2, 8, vec![2], 1), Metric::L2).unwrap();
+
+        assert_valid_partition(&leaves, 80, 8, 1);
+    }
+
+    #[test]
+    fn duplicate_points_return_iteration_limit_instead_of_oversized_leaf() {
+        let data = Matrix::new(1.0f32, 24, 4);
+        let error = partition(data.as_view(), config(2, 4, vec![1], 1), Metric::L2).unwrap_err();
+        let error = error.downcast::<PartitionError>().unwrap();
+
+        assert!(matches!(
+            error,
+            PartitionError::IterationLimit {
+                size: 24,
+                limit: MAX_PARTITION_ITERATIONS,
+                ..
+            }
+        ));
+    }
+
+    #[test]
+    fn global_merge_canonicalizes_small_leaf_membership() {
+        let leaves = vec![vec![9, 3, 1], vec![3, 2], vec![8]];
+
+        let merged = global_merge_small(leaves, 4, 8).unwrap();
+
+        assert_eq!(merged, vec![vec![1, 2, 3, 8, 9]]);
+    }
+
+    #[test]
+    fn global_merge_never_overfills_before_reaching_c_min() {
+        let leaves = vec![vec![0, 1, 2, 3], vec![4, 5, 6, 7], vec![8, 9, 10, 11]];
+
+        let merged = global_merge_small(leaves, 11, 11).unwrap();
+
+        assert_eq!(
+            merged,
+            vec![vec![0, 1, 2, 3, 4, 5, 6, 7], vec![8, 9, 10, 11]]
+        );
+    }
+
+    #[test]
+    fn global_merge_fills_exact_capacity_before_flushing() {
+        let merged = global_merge_small(vec![vec![0, 1], vec![2, 3]], 4, 4).unwrap();
+
+        assert_eq!(merged, vec![vec![0, 1, 2, 3]]);
+    }
+
+    #[test]
+    fn replicas_cover_every_point_once_or_more_per_replica() {
+        let data = directional_data(72, 5);
+        let leaves = partition(
+            data.as_view(),
+            config(3, 12, vec![3, 2], 3),
+            Metric::CosineNormalized,
+        )
+        .unwrap();
+
+        assert_valid_partition(&leaves, 72, 12, 3);
+    }
+
+    fn assert_partition_conversion_matches_f32<T>(label: &str, convert: impl Fn(u8) -> T)
+    where
+        T: crate::utils::VectorRepr + Send + Sync,
+    {
+        let points = 64;
+        // Partition gathering converts source vectors before GEMM. Exercise conversion
+        // tails around 4-, 8-, and 16-element boundaries and a second 16-lane chunk.
+        for dimensions in [1, 3, 4, 7, 8, 9, 15, 16, 17, 31, 32, 33] {
+            let raw: Vec<u8> = (0..points * dimensions)
+                .map(|index| {
+                    let point = index / dimensions;
+                    let dimension = index % dimensions;
+                    ((point * 5 + dimension * 7 + point * dimension) % 23) as u8
+                })
+                .collect();
+            let f32_data: Vec<f32> = raw.iter().map(|&value| value as f32).collect();
+            let converted: Vec<T> = raw.iter().copied().map(&convert).collect();
+            let config = config(2, 16, vec![2, 1], 1);
+            let expected = partition(
+                MatrixView::try_from(&f32_data, points, dimensions).unwrap(),
+                config.clone(),
+                Metric::L2,
+            )
+            .unwrap();
+            let actual = partition(
+                MatrixView::try_from(&converted, points, dimensions).unwrap(),
+                config,
+                Metric::L2,
+            )
+            .unwrap_or_else(|error| panic!("{label} dimensions={dimensions}: {error}"));
+
+            assert_valid_partition(&actual, points, 16, 1);
+            assert_eq!(
+                sorted_memberships(&actual),
+                sorted_memberships(&expected),
+                "{label} dimensions={dimensions}"
+            );
+        }
+    }
+
+    #[test]
+    fn f16_partition_matches_f32_across_dimension_boundaries() {
+        assert_partition_conversion_matches_f32("f16", |value| Half::from_f32(value as f32));
+    }
+
+    #[test]
+    fn u8_partition_matches_f32_across_dimension_boundaries() {
+        assert_partition_conversion_matches_f32("u8", |value| value);
+    }
+
+    #[test]
+    fn i8_partition_matches_f32_across_dimension_boundaries() {
+        // The same translation in every coordinate preserves L2 ordering.
+        assert_partition_conversion_matches_f32("i8", |value| value as i8 - 11);
+    }
+
+    #[test]
+    fn l2_leader_norms_preserve_scalar_reduction_order() {
+        fn next(state: &mut u64) -> f32 {
+            *state ^= *state << 13;
+            *state ^= *state >> 7;
+            *state ^= *state << 17;
+            (((*state >> 40) as f32 / 8_388_608.0) - 1.0) * 1_000.0
+        }
+
+        // This fixed case sits on opposite sides of the top-1 boundary depending
+        // on whether leader norms use the original scalar reduction or a SIMD
+        // reassociation. Point/leader dot products still go through the production
+        // GEMM; only the setup norm calculation is under test.
+        let dimensions = 129;
+        let mut state = 0x3a85_f952_c718_6e49;
+        let point: Vec<f32> = (0..dimensions).map(|_| next(&mut state)).collect();
+        let leader_zero: Vec<f32> = (0..dimensions).map(|_| next(&mut state)).collect();
+        let leader_one: Vec<f32> = (0..dimensions).map(|_| next(&mut state)).collect();
+        let data: Vec<f32> = leader_zero
+            .into_iter()
+            .chain(leader_one)
+            .chain(point)
+            .collect();
+        let data = MatrixView::try_from(data.as_slice(), 3, dimensions).unwrap();
+
+        let clusters = assign_to_leaders(
+            data,
+            &[2],
+            &[0, 1],
+            1,
+            Metric::L2,
+            &PartitionKernel::new(Metric::L2),
+            &StripeBufferPool::new((), 0, None),
+        )
+        .unwrap();
+
+        assert_eq!(clusters, [vec![], vec![2]]);
+    }
+
+    #[test]
+    fn all_metrics_produce_valid_partitions() {
+        let data = directional_data(64, 8);
+        let config = config(2, 20, vec![2], 1);
+
+        for metric in [
+            Metric::L2,
+            Metric::Cosine,
+            Metric::CosineNormalized,
+            Metric::InnerProduct,
+        ] {
+            let leaves = partition(data.as_view(), config.clone(), metric).unwrap();
+            assert_valid_partition(&leaves, 64, 20, 1);
+        }
+    }
+
+    #[test]
+    fn leader_count_is_bounded() {
+        assert_eq!(sample_num_leaders(1, 1.0), 1);
+        assert_eq!(sample_num_leaders(10, 0.01), 2);
+        assert_eq!(sample_num_leaders(50_000, 1.0), LEADER_CAP);
+    }
+
+    #[test]
+    fn replica_seed_derivation_is_stable_and_distinct() {
+        assert_eq!(replica_seed(0), 1_000);
+        assert_eq!(replica_seed(1), 8_919);
+    }
+
+    #[test]
+    fn assignment_stripes_use_power_of_two_point_counts() {
+        assert_eq!(assignment_stripe_point_count(1_000), 128);
+        assert_eq!(assignment_stripe_point_count(256), 512);
+        assert_eq!(
+            assignment_stripe_point_count(1),
+            MAX_ASSIGNMENT_STRIPE_POINTS
+        );
+    }
+
+    #[test]
+    fn stripe_buffer_pool_reuses_returned_capacity() {
+        let pool = StripeBufferPool::new((), 0, None);
+        let points = {
+            let mut buffers = pool.get_ref(());
+            buffers.points.resize(16, 0.0);
+            buffers.points.as_ptr()
+        };
+
+        let buffers = pool.get_ref(());
+        assert_eq!(buffers.points.as_ptr(), points);
+        assert_eq!(buffers.points.len(), 16);
+    }
+
+    #[test]
+    fn leader_assignment_handles_multiple_stripes() {
+        let points = 2_048;
+        let data: Vec<f32> = (0..points).map(|point| point as f32).collect();
+        let data = MatrixView::try_from(data.as_slice(), points, 1).unwrap();
+        let point_ids: Vec<u32> = (0..points as u32).collect();
+
+        let clusters = assign_to_leaders(
+            data,
+            &point_ids,
+            &[0, 2_047],
+            1,
+            Metric::L2,
+            &PartitionKernel::new(Metric::L2),
+            &StripeBufferPool::new((), 0, None),
+        )
+        .unwrap();
+
+        assert_eq!(clusters[0], (0..1_024).collect::<Vec<_>>());
+        assert_eq!(clusters[1], (1_024..2_048).collect::<Vec<_>>());
+    }
+
+    #[test]
+    fn parallel_scatter_matches_serial_order() {
+        let points: Vec<u32> = (0..PARALLEL_SCATTER_MIN_POINTS as u32).collect();
+        let assignments: Vec<u32> = points
+            .iter()
+            .flat_map(|point| [point % 7, (point + 3) % 7])
+            .collect();
+
+        let expected = scatter_serial(&points, &assignments, 2, 7).unwrap();
+        let actual = scatter_assignments(&points, &assignments, 2, 7).unwrap();
+
+        assert_eq!(actual, expected);
+    }
+
+    #[test]
+    fn rejects_empty_dataset() {
+        let data = Matrix::<f32>::new(0.0, 0, 4);
+        let error = partition(data.as_view(), config(1, 4, vec![1], 1), Metric::L2).unwrap_err();
+
+        assert_eq!(
+            error.downcast::<PartitionError>().unwrap(),
+            PartitionError::EmptyDataset
+        );
+    }
+
+    #[test]
+    fn rejects_zero_dimensions() {
+        let data = Matrix::<f32>::new(0.0, 4, 0);
+        let error = partition(data.as_view(), config(1, 4, vec![1], 1), Metric::L2).unwrap_err();
+
+        assert_eq!(
+            error.downcast::<PartitionError>().unwrap(),
+            PartitionError::EmptyDimensions
+        );
+    }
+
+    #[test]
+    fn rejects_invalid_gather_output_length() {
+        let data = Matrix::<f32>::new(0.0, 2, 2);
+        let error = gather_vectors(data.as_view(), &[0, 1], &mut [0.0; 3]).unwrap_err();
+
+        assert_eq!(
+            error.downcast::<PartitionError>().unwrap(),
+            PartitionError::InvalidBufferLength {
+                buffer: "gather output",
+                expected: 4,
+                actual: 3,
+            }
+        );
+    }
+
+    #[test]
+    fn rejects_assignment_to_an_unknown_leader() {
+        let error = scatter_serial(&[7], &[2], 1, 2).unwrap_err();
+
+        assert_eq!(
+            error.downcast::<PartitionError>().unwrap(),
+            PartitionError::InvalidBufferLength {
+                buffer: "leader assignment",
+                expected: 2,
+                actual: 3,
+            }
+        );
+    }
+
+    #[test]
+    fn rejects_empty_and_oversized_leaves() {
+        for (leaves, size) in [(vec![vec![]], 0), (vec![vec![0, 1, 2]], 3)] {
+            let error = validate_leaves(&leaves, 2).unwrap_err();
+            assert_eq!(
+                error.downcast::<PartitionError>().unwrap(),
+                PartitionError::InvalidLeaf { size, limit: 2 }
+            );
+        }
+    }
+}
